refactoring classification as mod
